--- a/src/components/LHNOptionsList/LHNOptionsList.js
+++ b/src/components/LHNOptionsList/LHNOptionsList.js
@@ -4,8 +4,8 @@ import {FlatList, View} from 'react-native';
 import _ from 'underscore';
 import CONST from '../../CONST';
 import styles from '../../styles/styles';
-import OptionRowLHNData from './OptionRowLHNData';
 import variables from '../../styles/variables';
+import OptionRowLHNDataWithFocus from './OptionRowLHNDataWithFocus';
 
 const propTypes = {
     /** Wrapper style for the section list */
@@ -63,7 +63,7 @@ function LHNOptionsList({style, contentContainerStyles, data, onSelectRow, optio
      * @return {Component}
      */
     const renderItem = ({item}) => (
-        <OptionRowLHNData
+        <OptionRowLHNDataWithFocus
             reportID={item}
             viewMode={optionMode}
             shouldDisableFocusOptions={shouldDisableFocusOptions}