--- a/src/components/LHNOptionsList/OptionRowLHNData.js
+++ b/src/components/LHNOptionsList/OptionRowLHNData.js
@@ -8,7 +8,7 @@ import {withReportCommentDrafts} from '../OnyxProvider';
 import SidebarUtils from '../../libs/SidebarUtils';
 import compose from '../../libs/compose';
 import ONYXKEYS from '../../ONYXKEYS';
-import withCurrentReportID, {withCurrentReportIDPropTypes, withCurrentReportIDDefaultProps} from '../withCurrentReportID';
+import withCurrentReportID from '../withCurrentReportID';
 import OptionRowLHN, {propTypes as basePropTypes, defaultProps as baseDefaultProps} from './OptionRowLHN';
 import * as Report from '../../libs/actions/Report';
 import * as UserUtils from '../../libs/UserUtils';
@@ -20,8 +20,8 @@ import CONST from '../../CONST';
 import reportActionPropTypes from '../../pages/home/report/reportActionPropTypes';
 
 const propTypes = {
-    /** If true will disable ever setting the OptionRowLHN to focused */
-    shouldDisableFocusOptions: PropTypes.bool,
+    /** If focused or not */
+    isFocused: PropTypes.bool,
 
     /** List of users' personal details */
     personalDetails: PropTypes.objectOf(participantPropTypes),
@@ -51,20 +51,17 @@ const propTypes = {
         /** The ID of the transaction */
         transactionID: PropTypes.string,
     }),
-
-    ...withCurrentReportIDPropTypes,
     ...basePropTypes,
 };
 
 const defaultProps = {
-    shouldDisableFocusOptions: false,
+    isFocused: false,
     personalDetails: {},
     fullReport: {},
     policy: {},
     parentReportActions: {},
     transaction: {},
     preferredLocale: CONST.LOCALES.DEFAULT,
-    ...withCurrentReportIDDefaultProps,
     ...baseDefaultProps,
 };
 
@@ -75,8 +72,7 @@ const defaultProps = {
  * re-render if the data really changed.
  */
 function OptionRowLHNData({
-    shouldDisableFocusOptions,
-    currentReportID,
+    isFocused,
     fullReport,
     reportActions,
     personalDetails,
@@ -91,7 +87,6 @@ function OptionRowLHNData({
     const reportID = propsToForward.reportID;
     // We only want to pass a boolean to the memoized component,
     // instead of a changing number (so we prevent unnecessary re-renders).
-    const isFocused = !shouldDisableFocusOptions && currentReportID === reportID;
 
     const parentReportAction = parentReportActions[fullReport.parentReportActionID];
 
@@ -170,9 +165,8 @@ const personalDetailsSelector = (personalDetails) =>
  * Thats also why the React.memo is used on the outer component here, as we just
  * use it to prevent re-renders from parent re-renders.
  */
-export default React.memo(
+const OptionRowLHNDataComposed = React.memo(
     compose(
-        withCurrentReportID,
         withReportCommentDrafts({
             propName: 'comment',
             transformValue: (drafts, props) => {
@@ -220,3 +214,15 @@ export default React.memo(
         }),
     )(OptionRowLHNData),
 );
+
+function OptionRowLHNDataWrapper({currentReportID, shouldDisableFocusOptions, ...props}) {
+    const isFocused = !shouldDisableFocusOptions && currentReportID === props.reportID;
+    return (
+        <OptionRowLHNDataComposed
+            {...props}
+            isFocused={isFocused}
+        />
+    );
+}
+
+export default React.memo(withCurrentReportID(OptionRowLHNDataWrapper));