--- a/src/components/LHNOptionsList/OptionRowLHNData.js
+++ b/src/components/LHNOptionsList/OptionRowLHNData.js
@@ -8,7 +8,7 @@ import {withReportCommentDrafts} from '../OnyxProvider';
 import SidebarUtils from '../../libs/SidebarUtils';
 import compose from '../../libs/compose';
 import ONYXKEYS from '../../ONYXKEYS';
-import withCurrentReportID from '../withCurrentReportID';
+import withCurrentReportID, {withCurrentReportIDPropTypes, withCurrentReportIDDefaultProps} from '../withCurrentReportID';
 import OptionRowLHN, {propTypes as basePropTypes, defaultProps as baseDefaultProps} from './OptionRowLHN';
 import * as Report from '../../libs/actions/Report';
 import * as UserUtils from '../../libs/UserUtils';
@@ -20,8 +20,8 @@ import CONST from '../../CONST';
 import reportActionPropTypes from '../../pages/home/report/reportActionPropTypes';
 
 const propTypes = {
-    /** If focused or not */
-    isFocused: PropTypes.bool,
+    /** If true will disable ever setting the OptionRowLHN to focused */
+    shouldDisableFocusOptions: PropTypes.bool,
 
     /** List of users' personal details */
     personalDetails: PropTypes.objectOf(participantPropTypes),
@@ -51,17 +51,20 @@ const propTypes = {
         /** The ID of the transaction */
         transactionID: PropTypes.string,
     }),
+
+    ...withCurrentReportIDPropTypes,
     ...basePropTypes,
 };
 
 const defaultProps = {
-    isFocused: false,
+    shouldDisableFocusOptions: false,
     personalDetails: {},
     fullReport: {},
     policy: {},
     parentReportActions: {},
     transaction: {},
     preferredLocale: CONST.LOCALES.DEFAULT,
+    ...withCurrentReportIDDefaultProps,
     ...baseDefaultProps,
 };
 
@@ -72,7 +75,8 @@ const defaultProps = {
  * re-render if the data really changed.
  */
 function OptionRowLHNData({
-    isFocused,
+    shouldDisableFocusOptions,
+    currentReportID,
     fullReport,
     reportActions,
     personalDetails,
@@ -87,6 +91,7 @@ function OptionRowLHNData({
     const reportID = propsToForward.reportID;
     // We only want to pass a boolean to the memoized component,
     // instead of a changing number (so we prevent unnecessary re-renders).
+    const isFocused = !shouldDisableFocusOptions && currentReportID === reportID;
 
     const parentReportAction = parentReportActions[fullReport.parentReportActionID];
 
@@ -165,8 +170,9 @@ const personalDetailsSelector = (personalDetails) =>
  * Thats also why the React.memo is used on the outer component here, as we just
  * use it to prevent re-renders from parent re-renders.
  */
-const OptionRowLHNDataComposed = React.memo(
+export default React.memo(
     compose(
+        withCurrentReportID,
         withReportCommentDrafts({
             propName: 'comment',
             transformValue: (drafts, props) => {
@@ -214,15 +220,3 @@ const OptionRowLHNDataComposed = React.memo(
         }),
     )(OptionRowLHNData),
 );
-
-function OptionRowLHNDataWrapper({currentReportID, shouldDisableFocusOptions, ...props}) {
-    const isFocused = !shouldDisableFocusOptions && currentReportID === props.reportID;
-    return (
-        <OptionRowLHNDataComposed
-            {...props}
-            isFocused={isFocused}
-        />
-    );
-}
-
-export default React.memo(withCurrentReportID(OptionRowLHNDataWrapper));