--- a/src/components/LHNOptionsList/OptionRowLHNDataWithFocus.js
+++ b/src/components/LHNOptionsList/OptionRowLHNDataWithFocus.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import withCurrentReportID, {withCurrentReportIDPropTypes, withCurrentReportIDDefaultProps} from '../withCurrentReportID';
+import OptionRowLHNData from './OptionRowLHNData';
+
+const propTypes = {
+    ...withCurrentReportIDPropTypes,
+    shouldDisableFocusOptions: PropTypes.bool,
+};
+
+const defaultProps = {
+    ...withCurrentReportIDDefaultProps,
+    shouldDisableFocusOptions: false,
+};
+
+function OptionRowLHNDataWithFocus({currentReportID, shouldDisableFocusOptions, ...props}) {
+    const isFocused = !shouldDisableFocusOptions && currentReportID === props.reportID;
+
+    return (
+        <OptionRowLHNData
+            {...props}
+            isFocused={isFocused}
+        />
+    );
+}
+
+OptionRowLHNDataWithFocus.defaultProps = defaultProps;
+OptionRowLHNDataWithFocus.propTypes = propTypes;
+OptionRowLHNDataWithFocus.displayName = 'OptionRowLHNDataWithFocus';
+
+export default withCurrentReportID(OptionRowLHNDataWithFocus);