--- a/src/components/LHNOptionsList/OptionRowLHNDataWithFocus.js
+++ b/src/components/LHNOptionsList/OptionRowLHNDataWithFocus.js
@@ -18,6 +18,7 @@ function OptionRowLHNDataWithFocus({currentReportID, shouldDisableFocusOptions, 
 
     return (
         <OptionRowLHNData
+            // eslint-disable-next-line react/jsx-props-no-spreading
             {...props}
             isFocused={isFocused}
         />