## Intro

Transition between chats (reports), seems to be quite laggy. In this document, we try to investigate the issue and come up with possible solution. 

## Steps to reproduce

1. Load two chats
2. Navigate from one to another

## Recording

https://github.com/Expensify/App/assets/62747088/85bb45e8-e394-4f61-bc8a-5fe57361ef77

## Investigations

It looks like there is a big overhead in form of rendering LHN’s list of `OptionRowLHNData` components.
<img width="1502" alt="BeforeFixes" src="https://github.com/Expensify/App/assets/62747088/9a9f8987-52d8-4d0e-b669-eaf1193f3770">

In the profiling trace we see that every item in the list is re-rendered. Each re-render takes up to **5ms** to complete, which with current list size can significantly throttle rendering.

The re-rendering is caused by change in `currentReportID` state, which is provided by `withCurrentReportID` HOC. `currentReportID` prop is used in  `OptionRowLHNData` in one place, to find out wether component should be focused or not:

```tsx
const isFocused = !shouldDisableFocusOptions && currentReportID === reportID;
```

The `isFocused` is then directly passed down into child.

The change in `currentReportID` breaks memoization, causing heavy re-render of each item.

## Solutions

### Wrapper component

As `currentReportID` is only used **********once********** and `isFocused` is a boolean flag, we can extract this logic into separate **************wrapper************** component and refactor `OptionRowLHNData` to receive primitive `isFocused` prop, which will not break memoization. 

<img width="1502" alt="AfterFixes copy" src="https://github.com/Expensify/App/assets/62747088/eb03b8e8-9c01-472a-ad6c-117342323d8d">

By applying these changes, we were able to shrink down the rendering time of List component to **~100ms**, which will be even more significant as number of list items grows. The rest of rendering overhead seems to be related to different components, `Report` itself etc.

The rendering time should be now only merely related to list size, as it only triggers one heavy re-render.

Wrapper component would look more or less like this:

```tsx
function OptionRowLHNDataWrapper({currentReportID, shouldDisableFocusOptions, ...props}) {
    const isFocused = !shouldDisableFocusOptions && currentReportID === props.reportID;
    return (
        <OptionRowLHNDataComposed
            {...props}
            isFocused={isFocused}
        />
    );
}
```

This solution does not add anything beyond a simple wrapper so it is quite easy to do and should not be error-prone.

### Use `withCurrentReportID` in list component (Alternative)

<img width="1502" alt="AfterFixesDetailedAlternatives copy" src="https://github.com/Expensify/App/assets/62747088/3353375d-47b0-4151-8f93-8120d036a1b5">

Alternatively, we can use `withCurrentReportID` once at the level of list component instead of in each children. Then we will directly pass `isFocused` prop into children, so it won’t break memoization of list items.

The solution is more elegant, as it uses `currentReportID` provider once and does not expose such internals into children components. On the flipside, we need to re-render list itself, which might add some overhead.

In this scenario, it takes about **********~120ms********** to render the list, which is about **********15ms********** slower than previous solution.

## Conclusions

While we managed to reduce rendering times, we still need to think on which solution should be implemented. Furthermore, it might be a good idea to investigate rendering of list items, as it is quite simple component but still takes ************~5ms************ to render.