@kacper-mikolajczak 2023-10-02
Hi @muttmuure, here is a draft of partial improvements that I've done recently. I wanted to share it on Slack but it is quite hard to put such document into Slack message :D Let me know what you think about results and overall shape of such document. In the meantime I will work on further improvements