@mountiny 2023-10-02
@kacper-mikolajczak Thanks! I think this looks great!

I agree that the second solution leads to more readable code, adding that one wrapper feels a bit odd to me.

@kacper-mikolajczak could you summarize these two options and post it to #newdot-performance room? I think it would be great to get opinions of other agency experts and engineers on this and then we can implement this. Thanks!