@kacper-mikolajczak 2023-10-02
@mountiny  Yeah definitely! Thanks for the feedback 👍