@kacper-mikolajczak 2023-10-04
BTW, on iPhone 13 simulator the difference is huge

<img src="https://github.com/Expensify/App/assets/62747088/8e37e76b-d075-4f40-8e50-85f4eb735aa0" width=300 />
<img src="https://github.com/Expensify/App/assets/62747088/5e1538b4-1527-4c1f-a8ef-ec9c161ac365" width=300 />