@kacper-mikolajczak 2023-10-05
@mountiny Yeah that does look promising! Wonder what would be a perceived perf gain, as the list is out of the view when the render happens (it is covered by report screen). But hey, at least we reduced the resources consumption a lil bit :D