@kacper-mikolajczak 2023-10-05
Sure @mountiny, testing it right now!