@kacper-mikolajczak 2023-10-05
Here are the results @mountiny.

### Mobile (iOS)

From testing on iOS it seems that the app performance is highly coupled to the amount of chats present. It is no surprise that it is influencing performance but the app should not slow down almost exponentially when amount of chats grows.

My first impression is that there are some heavy calculations based on entire data set going on in each item. I am going to investigate it further in some future, as it is out of the scope of the current RP.

Main branch:
<img width="450" alt="Screenshot 2023-10-05 at 11 39 15" src="https://github.com/Expensify/App/assets/62747088/72a17b89-aa8e-4a10-99c9-4323b769174b">
Current branch:
<img width="450" alt="Screenshot 2023-10-05 at 11 38 40" src="https://github.com/Expensify/App/assets/62747088/cfe363dc-51d8-43ee-b5ca-82273d0dd598">

### Web
For web, I was not able to load the app at all - no errors or anything particularly wrong. It after inspecting devtools it seems that there are a lot of calculations (probably related to what Matt and Monil discussed on Slack).