@mountiny 2023-10-05
@allroundexperts Thank you! Do you want to raise that here? https://expensify.slack.com/archives/C05LX9D6E07/p1696354046851689 it might be quicker to discuss and more people ot bound the ideas off