@kacper-mikolajczak 2023-10-06
Hi @allroundexperts, thanks for your feedback! 👍 
I agree with your concern about growing complexity and while I was thinking about the solution you've suggested, I came to the conclusion that it would eventually lead us to adding more abstraction that we might need at the moment.

In the current approach, we simply couple the OptionRowLHNData with the tiny wrapper in the form of OptionRowLHNDataWithFocus (naming itself is another topic - I am not sure it's correct). That's the whole abstraction we introduce.

In the suggested approach, we would first need to create withCurrentReportFocused HOC, but then we would still need to bound it to the OptionRowLHNData, which will lead us to creating OptionRowLHNDataWithFocus for this purpose. We just simply moved part of the logic (as shouldDisableFocusOptions still needs to be handled inside the wrapper component).

I agree that tightly coupling should be less preferred than composition, but the question is how much do we need to have withCurrentReportFocused abstraction as it is potentially just one line of extracted logic that seems kind of specific. Plus, it looks like it's going to be only used in one place and by sticking to YAGNI rule, we can think about leaving it as-is, wdyt?

Having that said, I am not as familiar with the codebase as you are, please let me know if I understood it right. Thanks a lot for the discussion :smile: