@OSBotify 2023-10-09
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/6462270103) to staging by https://github.com/mountiny in version: 1.3.80-0 🚀

platform | result
---|---
🤖 android 🤖|success ✅
🖥 desktop 🖥|success ✅
🍎 iOS 🍎|success ✅
🕸 web 🕸|success ✅