@OSBotify 2023-10-11
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/6477977692) to production by https://github.com/jasperhuangg in version: 1.3.80-3 🚀

platform | result
---|---
🤖 android 🤖|success ✅
🖥 desktop 🖥|success ✅
🍎 iOS 🍎|failure ❌
🕸 web 🕸|success ✅