@OSBotify 2023-10-13
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/6513766880) to production by https://github.com/jasperhuangg in version: 1.3.83-11 🚀

platform | result
---|---
🤖 android 🤖|failure ❌
🖥 desktop 🖥|success ✅
🍎 iOS 🍎|failure ❌
🕸 web 🕸|success ✅