@OSBotify 2023-10-16
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/6539292081) to production by https://github.com/francoisl in version: 1.3.84-10 🚀

platform | result
---|---
🤖 android 🤖|skipped 🚫
🖥 desktop 🖥|skipped 🚫
🍎 iOS 🍎|skipped 🚫
🕸 web 🕸|skipped 🚫