@OSBotify 2023-10-16
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/6539526963) to production by https://github.com/francoisl in version: 1.3.84-10 🚀

platform | result
---|---
🤖 android 🤖|success ✅
🖥 desktop 🖥|success ✅
🍎 iOS 🍎|failure ❌
🕸 web 🕸|success ✅